+++ imio/dms/mail/dmsmail.py
@@ -227,8 +227,11 @@ class IMView(DmsDocumentView):
         View form redefinition to customize fields.
     """
 
-    def updateWidgets(self):
+    def updateWidgets(self, prefix=None):
         super(IMView, self).updateWidgets()
+        # this is added to escape treatment when displaying single widget in column
+        if prefix == 'escape':
+            return
         for field in ['ITask.assigned_group', 'ITask.enquirer']:
             self.widgets[field].mode = HIDDEN_MODE